# backend-spring/HealthCARE-App-mrx-part4/swagger-DOCUMENTATION.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "HealthCare+ API",
                 "description":  "Documentation officielle de l\u0027API HealthCare+ : authentification JWT, gestion des patients, médecins, rendez-vous et dossiers médicaux.",
                 "contact":  {
                                 "name":  "HealthCare+",
                                 "email":  "support@healthcare.local"
                             },
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  []
                     }
                 ],
...
    "paths":  {
...
                  "/dossiers/{id}":  {
                                         "get":  {
                                                     "tags":  [
                                                                  "DossierMedical"
                                                              ],
                                                     "operationId":  "get_3",
                                                     "parameters":  [
                                                                        {
                                                                            "name":  "id",
                                                                            "in":  "path",
                                                                            "required":  true,
                                                                            "schema":  {
                                                                                           "type":  "integer",
                                                                                           "format":  "int32"
                                                                                       }
                                                                        }
                                                                    ],
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK",
                                                                                   "content":  {
                                                                                                   "*/*":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/DossierMedicalGetDTO"
                                                                                                                          }
                                                                                                           }
                                                                                               }
                                                                               }
                                                                   },
                                                     "security":  [
                                                                      {
                                                                          "bearerAuth":  [

                                                                                         ]
                                                                      }
                                                                  ]
                                                 },
                                         "put":  {
                                                     "tags":  [
                                                                  "DossierMedical"
                                                              ],
                                                     "operationId":  "update_4",
                                                     "parameters":  [
                                                                        {
                                                                            "name":  "id",
                                                                            "in":  "path",
                                                                            "required":  true,
                                                                            "schema":  {
                                                                                           "type":  "integer",
                                                                                           "format":  "int32"
                                                                                       }
                                                                        }
                                                                    ],
                                                     "requestBody":  {
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/DossierMedicalUpdateDTO"
                                                                                                                             }
                                                                                                              }
                                                                                     },
                                                                         "required":  true
                                                                     },
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK",
                                                                                   "content":  {
                                                                                                   "*/*":  {
                                                                                                               "schema":  {
                                                                                                                              "$ref":  "#/components/schemas/DossierMedicalGetDTO"
                                                                                                                          }
                                                                                                           }
                                                                                               }
                                                                               }
                                                                   },
                                                     "security":  [
                                                                      {
                                                                          "bearerAuth":  [

                                                                                         ]
                                                                      }
                                                                  ]
                                                 },
                                         "delete":  {
                                                        "tags":  [
                                                                     "DossierMedical"
                                                                 ],
                                                        "operationId":  "delete_4",
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "id",
                                                                               "in":  "path",
                                                                               "required":  true,
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "format":  "int32"
                                                                                          }
                                                                           }
                                                                       ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "OK"
                                                                                  }
                                                                      },
                                                        "security":  [
                                                                         {
                                                                             "bearerAuth":  [

                                                                                            ]
                                                                         }
                                                                     ]
                                                    }
                                     },
...
                  "/dossiers":  {
                                    "get":  {
                                                "tags":  [
                                                             "DossierMedical"
                                                         ],
                                                "operationId":  "list_4",
                                                "parameters":  [
                                                                   {
                                                                       "name":  "page",
                                                                       "in":  "query",
                                                                       "required":  false,
                                                                       "schema":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int32",
                                                                                      "default":  0
                                                                                  }
                                                                   },
                                                                   {
                                                                       "name":  "size",
                                                                       "in":  "query",
                                                                       "required":  false,
                                                                       "schema":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int32",
                                                                                      "default":  5
                                                                                  }
                                                                   },
                                                                   {
                                                                       "name":  "sort",
                                                                       "in":  "query",
                                                                       "required":  false,
                                                                       "schema":  {
                                                                                      "type":  "string",
                                                                                      "default":  "dateCreation"
                                                                                  }
                                                                   },
                                                                   {
                                                                       "name":  "direction",
                                                                       "in":  "query",
                                                                       "required":  false,
                                                                       "schema":  {
                                                                                      "type":  "string",
                                                                                      "default":  "desc"
                                                                                  }
                                                                   }
                                                               ],
                                                "responses":  {
                                                                  "200":  {
                                                                              "description":  "OK",
                                                                              "content":  {
                                                                                              "*/*":  {
                                                                                                          "schema":  {
                                                                                                                         "$ref":  "#/components/schemas/PageDossierMedicalGetDTO"
                                                                                                                     }
                                                                                                      }
                                                                                          }
                                                                          }
                                                              },
                                                "security":  [
                                                                 {
                                                                     "bearerAuth":  [

                                                                                    ]
                                                                 }
                                                             ]
                                            },
                                    "post":  {
                                                 "tags":  [
                                                              "DossierMedical"
                                                          ],
                                                 "operationId":  "create_2",
                                                 "requestBody":  {
                                                                     "content":  {
                                                                                     "application/json":  {
                                                                                                              "schema":  {
                                                                                                                             "$ref":  "#/components/schemas/DossierMedicalPostDTO"
                                                                                                                         }
                                                                                                          }
                                                                                 },
                                                                     "required":  true
                                                                 },
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "OK",
                                                                               "content":  {
                                                                                               "*/*":  {
                                                                                                           "schema":  {
                                                                                                                          "$ref":  "#/components/schemas/DossierMedicalGetDTO"
                                                                                                                      }
                                                                                                       }
                                                                                           }
                                                                           }
                                                               },
                                                 "security":  [
                                                                  {
                                                                      "bearerAuth":  [

                                                                                     ]
                                                                  }
                                                              ]
                                             }
                                },
...
                  "/dossiers/me":  {
                                       "get":  {
                                                   "tags":  [
                                                                "DossierMedical"
                                                            ],
                                                   "operationId":  "getMyDossier",
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK",
                                                                                 "content":  {
                                                                                                 "*/*":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/DossierMedicalGetDTO"
                                                                                                                        }
                                                                                                         }
                                                                                             }
                                                                             }
                                                                 },
                                                   "security":  [
                                                                    {
                                                                        "bearerAuth":  [

                                                                                       ]
                                                                    }
                                                                ]
                                               }
                                   },
                  "/api/download/dossier-medical/{id}":  {
                                                             "get":  {
                                                                         "tags":  [
                                                                                      "FileDownload"
                                                                                  ],
                                                                         "operationId":  "downloadDossierMedical",
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "id",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "integer",
                                                                                                               "format":  "int32"
                                                                                                           }
                                                                                            }
                                                                                        ],
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "OK",
                                                                                                       "content":  {
                                                                                                                       "*/*":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "type":  "string",
                                                                                                                                                  "format":  "byte"
                                                                                                                                              }
                                                                                                                               }
                                                                                                                   }
                                                                                                   }
                                                                                       },
                                                                         "security":  [
                                                                                          {
                                                                                              "bearerAuth":  [

                                                                                                             ]
                                                                                          }
                                                                                      ]
                                                                     }
                                                         }
              },
    "components":  {
                       "schemas":  {
...
                                       "DossierMedicalUpdateDTO":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "diagnostic":  {
                                                                                                             "type":  "string",
                                                                                                             "minLength":  1
                                                                                                         },
                                                                                          "observation":  {
                                                                                                              "type":  "string",
                                                                                                              "minLength":  1
                                                                                                          }
                                                                                      },
                                                                       "required":  [
                                                                                        "diagnostic",
                                                                                        "observation"
                                                                                    ]
                                                                   },
                                       "DossierMedicalGetDTO":  {
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "id":  {
                                                                                                  "type":  "integer",
                                                                                                  "format":  "int32"
                                                                                              },
                                                                                       "diagnostic":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                       "observation":  {
                                                                                                           "type":  "string"
                                                                                                       },
                                                                                       "dateCreation":  {
                                                                                                            "type":  "string",
                                                                                                            "format":  "date"
                                                                                                        },
                                                                                       "patientNom":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                       "patientPrenom":  {
                                                                                                             "type":  "string"
                                                                                                         }
                                                                                   }
                                                                },
...
                                       "DossierMedicalPostDTO":  {
                                                                     "type":  "object",
                                                                     "properties":  {
                                                                                        "diagnostic":  {
                                                                                                           "type":  "string",
                                                                                                           "minLength":  1
                                                                                                       },
                                                                                        "observation":  {
                                                                                                            "type":  "string",
                                                                                                            "minLength":  1
                                                                                                        },
                                                                                        "dateCreation":  {
                                                                                                             "type":  "string",
                                                                                                             "format":  "date"
                                                                                                         },
                                                                                        "patientId":  {
                                                                                                          "type":  "integer",
                                                                                                          "format":  "int32"
                                                                                                      }
                                                                                    },
                                                                     "required":  [
                                                                                      "dateCreation",
                                                                                      "diagnostic",
                                                                                      "observation",
                                                                                      "patientId"
                                                                                  ]
                                                                 },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "name":  "bearerAuth",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}